adds better scoring decisions

+++ Arena/ArenaOnlineGameModes/TeamBattle/TeamBattle.cs
@@ -267,6 +267,14 @@ namespace RainMeadow.Arena.ArenaOnlineGameModes.TeamBattle
 
                         if (aIsWinningTeam && bIsWinningTeam)
                         {
+                            if (A.allKills.Count != B.allKills.Count)
+                            {
+                                return A.allKills.Count > B.allKills.Count;
+                            }
+                            else if (A.deaths != B.deaths)
+                            {
+                                return A.deaths > B.deaths;
+                            }
                             return A.score > B.score; // If both are on winning team, sort by kill value
                         }
                     }
